fix inner for loop on large inputs, remove xorBytes function
for large inputs where d > 16, maxJ becomes >1, which causes the code
to enter the inner for loop. previously there was an issue where it an
input could be encrypted then the decrypted form didn’t match the
original plaintext. this commit fixes that.

note that correctness of the cipher text in these scenarios is still
unknown since no test vector covers it